Adds content column to documents and revisions to store current JSON snapshot
