+++ server/emails/templates/CommentCreatedEmail.tsx
@@ -2,10 +2,10 @@ import * as React from "react";
 import { NotificationEventType } from "@shared/types";
 import { Day } from "@shared/utils/time";
 import { Collection, Comment, Document } from "@server/models";
-import DocumentHelper from "@server/models/helpers/DocumentHelper";
 import HTMLHelper from "@server/models/helpers/HTMLHelper";
 import NotificationSettingsHelper from "@server/models/helpers/NotificationSettingsHelper";
 import ProsemirrorHelper from "@server/models/helpers/ProsemirrorHelper";
+import TextHelper from "@server/models/helpers/TextHelper";
 import BaseEmail, { EmailProps } from "./BaseEmail";
 import Body from "./components/Body";
 import Button from "./components/Button";
@@ -74,7 +74,7 @@ export default class CommentCreatedEmail extends BaseEmail<
       }
     );
 
-    content = await DocumentHelper.attachmentsToSignedUrls(
+    content = await TextHelper.attachmentsToSignedUrls(
       content,
       document.teamId,
       (4 * Day) / 1000
