+++ server/models/helpers/DocumentHelper.test.ts
@@ -1,5 +1,5 @@
 import Revision from "@server/models/Revision";
-import { buildDocument, buildUser } from "@server/test/factories";
+import { buildDocument } from "@server/test/factories";
 import DocumentHelper from "./DocumentHelper";
 
 describe("DocumentHelper", () => {
@@ -10,28 +10,6 @@ describe("DocumentHelper", () => {
 
   afterAll(() => {
     jest.useRealTimers();
-  });
-
-  describe("replaceTemplateVariables", () => {
-    it("should replace {time} with current time", async () => {
-      const user = await buildUser();
-      const result = DocumentHelper.replaceTemplateVariables(
-        "Hello {time}",
-        user
-      );
-
-      expect(result).toBe("Hello 12 00 AM");
-    });
-
-    it("should replace {date} with current date", async () => {
-      const user = await buildUser();
-      const result = DocumentHelper.replaceTemplateVariables(
-        "Hello {date}",
-        user
-      );
-
-      expect(result).toBe("Hello January 1 2021");
-    });
   });
 
   describe("parseMentions", () => {
