+++ server/presenters/revision.ts
@@ -1,6 +1,7 @@
 import parseTitle from "@shared/utils/parseTitle";
 import { traceFunction } from "@server/logging/tracing";
 import { Revision } from "@server/models";
+import DocumentHelper from "@server/models/helpers/DocumentHelper";
 import presentUser from "./user";
 
 async function presentRevision(revision: Revision, diff?: string) {
@@ -11,7 +12,7 @@ async function presentRevision(revision: Revision, diff?: string) {
     id: revision.id,
     documentId: revision.documentId,
     title: strippedTitle,
-    text: revision.text,
+    text: DocumentHelper.toMarkdown(revision),
     emoji: revision.emoji ?? emoji,
     html: diff,
     createdAt: revision.createdAt,
